Restore standalone integration
