+++ tests/integration/conftest.py
@@ -10,6 +10,21 @@ import pytest
 from pytest_operator.plugin import OpsTest
 
 
+@pytest.mark.abort_on_fail
+@pytest.fixture(scope="module")
+async def application_charm(ops_test: OpsTest):
+    """Build the application charm."""
+    test_charm_path = "./tests/integration/relations/pgbouncer_provider/application-charm"
+    return await ops_test.build_charm(test_charm_path)
+
+
+@pytest.mark.abort_on_fail
+@pytest.fixture(scope="module")
+async def pgb_charm(ops_test: OpsTest):
+    """Build the pgbouncer charm."""
+    return await ops_test.build_charm(".")
+
+
 @pytest.fixture(scope="module")
 def ops_test(ops_test: OpsTest) -> OpsTest:
     if os.environ.get("CI") == "true":
